arreglando funcion registrar
