+++ src/main.js
@@ -3,6 +3,7 @@
 //import { myFunction } from './lib/index.js';
 
 //myFunction();
+
 import { changeVistas } from "./controlador-vistas/controlador.js"
 
  export const inicializa = () => {
